+++ CloudShell.Persistence/Migrations/CloudShell/20260702164000_AddUsagePersistence.Designer.cs
@@ -3,6 +3,7 @@ using System;
 using CloudShell.Persistence;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
+using Microsoft.EntityFrameworkCore.Migrations;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 
 #nullable disable
@@ -10,14 +11,11 @@ using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 namespace CloudShell.Persistence.Migrations.CloudShell
 {
     [DbContext(typeof(CloudShellDbContext))]
-    partial class CloudShellDbContextModelSnapshot : ModelSnapshot
+    [Migration("20260702164000_AddUsagePersistence")]
+    partial class _20260702164000_AddUsagePersistence
     {
-        // If you encounter a merge conflict in the line below, it means you need to
-        // discard one of the migration branches and recreate its migrations on top of
-        // the other branch. See https://aka.ms/efcore-docs-migrations-conflicts for more info.
-        public override string LatestMigrationId => "20260702164000_AddUsagePersistence";
-
-        protected override void BuildModel(ModelBuilder modelBuilder)
+        /// <inheritdoc />
+        protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder.HasAnnotation("ProductVersion", "11.0.0-preview.4.26230.115");